api_register/requirements.txt
async-timeout==5.0.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
mariadb==1.1.10
MarkupSafe==3.0.2
packaging==24.1
pip==22.0.2
redis==5.2.0
setuptools==59.6.0
Werkzeug==3.1.1
wheel==0.37.1
gunicorn==20.1.0
python-dotenv